// client/src/store/member/member.types.ts
export enum ActionTypes {
  FETCH_MEMBERS_START = "FETCH_MEMBERS_START",
  FETCH_MEMBERS_SUCCESS = "FETCH_MEMBERS_SUCCESS",
  FETCH_MEMBERS_FAILURE = "FETCH_MEMBERS_FAILURE",
  ADD_MEMBER_SUCCESS = "ADD_MEMBER_SUCCESS",
  ADD_MEMBER_FAILURE = "ADD_MEMBER_FAILURE",
  DELETE_MEMBER_SUCCESS = "DELETE_MEMBER_SUCCESS",
  DELETE_MEMBER_FAILURE = "DELETE_MEMBER_FAILURE",
  EDIT_MEMBER_SUCCESS = "EDIT_MEMBER_SUCCESS",
  EDIT_MEMBER_FAILURE = "EDIT_MEMBER_FAILURE",
}